TASK: Refactor to middleware for Flow 7
