+++ Classes/Http/Middleware/MeasureServerTimingMiddleware.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace t3n\Neos\Debug\Http;
+namespace t3n\Neos\Debug\Http\Middleware;
 
 /**
  * This file is part of the t3n.Neos.Debugger package.
@@ -15,11 +15,13 @@ namespace t3n\Neos\Debug\Http;
  */
 
 use Neos\Flow\Annotations as Flow;
-use Neos\Flow\Http\Component\ComponentContext;
-use Neos\Flow\Http\Component\ComponentInterface;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 use t3n\Neos\Debug\Service\DebugService;
 
-class MeasureServerTimingComponent implements ComponentInterface
+class MeasureServerTimingMiddleware implements MiddlewareInterface
 {
     /**
      * @Flow\InjectConfiguration(path="serverTimingHeader.enabled")
@@ -35,15 +37,14 @@ class MeasureServerTimingComponent implements ComponentInterface
      */
     protected $debugService;
 
-    /**
-     * @inheritDoc
-     */
-    public function handle(ComponentContext $componentContext)
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
-        if (! $this->enabled) {
-            return;
+        $response = $handler->handle($request);
+
+        if ($this->enabled) {
+            $this->debugService->startRequestTimer();
         }
 
-        $this->debugService->startRequestTimer();
+        return $response;
     }
 }
