+++ composer.json
@@ -4,7 +4,7 @@
     "license": "MIT",
     "type": "neos-package",
     "require": {
-        "neos/neos": "~3.3.0 || ~4.0 || ~5.0"
+        "neos/neos": "^7.0"
     },
     "autoload": {
         "psr-4": {
